backend/pyproject.toml
[tool.poetry]
name = "beamng-worldforge-backend"
version = "0.1.0"
description = "Backend API for BeamNG.WorldForge - Automatic map generation from satellite data"
authors = ["BeamNG.WorldForge Team"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
earthengine-api = "^0.1.384"
rasterio = "^1.3.9"
numpy = "^1.26.3"
pillow = "^10.2.0"
httpx = "^0.26.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

